[Issue-1024] [Transaction] [SendFund] Update handle warning and error
