+++ packages/extension-koni-ui/src/Popup/Account/AccountExport.tsx
@@ -342,6 +342,7 @@ const Component: React.FC<Props> = (props: Props) => {
                       required: true
                     }
                   ]}
+                  statusHelpAsTooltip={true}
                 >
                   <Input
                     disabled={loading}
