+++ packages/extension-koni-ui/src/Popup/Settings/Chains/ChainDetail.tsx
@@ -444,6 +444,7 @@ function Component ({ className = '' }: Props): React.ReactElement<Props> {
                   <Form.Item
                     name={'blockExplorer'}
                     rules={[{ validator: blockExplorerValidator }]}
+                    statusHelpAsTooltip={true}
                   >
                     <Input placeholder={t('Block explorer')} />
                   </Form.Item>
@@ -458,6 +459,7 @@ function Component ({ className = '' }: Props): React.ReactElement<Props> {
                   <Form.Item
                     name={'crowdloanUrl'}
                     rules={[{ validator: crowdloanUrlValidator }]}
+                    statusHelpAsTooltip={true}
                   >
                     <Input placeholder={t('Crowdloan URL')} />
                   </Form.Item>
