+++ packages/extension-koni-ui/src/hooks/transaction/confirmation/useOpenDetailModal.ts
