+++ packages/extension-koni-ui/src/hooks/transaction/index.ts
@@ -1,0 +1,6 @@
+// Copyright 2019-2022 @polkadot/extension-koni-ui authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+export { default as useHandleSubmitTransaction } from './useHandleSubmitTransaction';
+
+export * from './confirmation';
